+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- ### Fixed -->
 <!-- ### Security -->
 
-## [Unreleased]
+## v0.5.1 (2018-12-19)
 ### Fixed
 - Fixed output when tags are styled to be underlined
 - Fixed `--check` output when you use `--user` to filter output